packages/numba-hip/numba-hip-requirements.txt
-i https://test.pypi.org/simple/
hip-python
hip-python-as-cuda
rocm-llvm-python
